sstables: no longer touch directory when writing SSTables
After the previous two patches, the CF directory where the SSTable will live is
guaranteed to always exist: system CFs' are touched at boot while newly created
tables' are touched when the creation mutations are announced.

With that in place, there is no more need for the recursive touch in the
SSTables path.

Signed-off-by: Glauber Costa <glommer@scylladb.com>