sstables: wire priority for write path
All visible variants of write_component now take an io_priority. It is set to
Seastar's default priority if nothing else is passed. We could mandate a
parameter because in practice the caller always want to specify a priority
during normal operation anyway. But we'll leave such default for the benefit of
tests.

Signed-off-by: Glauber Costa <glauber@scylladb.com>