add a priority class to mutation readers
SSTables already have a priority argument wired to their read path. However,
most of our reads do not call that interface directly, but employ the services
of a mutation reader instead.

Some of those readers will be used to read from a single place which is either
obviously not doing any I/O, or doing I/O always into a very specific context.
Those don't have to take any parameter. But those that can be used for multiple
purposes will need to be augmented with a parameter so we can specify the
priority class we want to use.

Right now, whenever we need to pass a class, we pass Seastar's default priority
class.

Signed-off-by: Glauber Costa <glauber@scylladb.com>