database: more accurate pending bytes tracking for memtables.
Our memtable pending flushes tracking is less accurate than what we would like:
specially given the fact that a complete, big, memtable may take some time to
write, we are left with the information that there is one full pending memtable
even when we're almost done writing it.

As we write partitions to stable storage, we can and should integrate that
information into our metrics. Since it is still useful to know how many
memtable operations are in flight, we will leave that metric alone, but add a
correction term to the metric that shows the number of bytes left to write.

Unfortunately, we don't really know how big a partition really is. Computing
that as we go is possible, but expensive. So we'll resort to a simple linear
averaging of the value: by querying how many partitions are already written, we
can use that information to estimate which percentage of the memtable have been
already written.

Signed-off-by: Glauber Costa <glauber@scylladb.com>