database: be less abrupt when trottling
Our database-based throttling mechanism allows us to go up to a certain point
of dirty memory and at that point, blocks abruptly, waiting for memory to be
released. That dirty memory is basically filled with memtables, and those memtables
may take a long while to be written back to back storage. This means that we stop the
database for a long time, even if we are already quite ahead in writing those memtables
to memory.

We can do better by factoring in the amount of memory already written.  We know
that this amount of memory has been flushed already, but that does not mean it
has been released. We can proactively use this to start letting requests go
because it is soon to be released, but we must be careful not to blow all
available memory.

Signed-off-by: Glauber Costa <glauber@scylladb.com>