sstables: wrap compressed file with a file_writer class as well
This is technically not needed, proof of that being that we have lived
without it so far. However, this creates an asymmetry in the sense that
the checksumm file handler is created through a file_writer object and
the compressed file is not.

In preparation for an upcoming patch where more information will be added
to file_writer, this patch makes the creation of the file handle homogeneous.

Signed-off-by: Glauber Costa <glauber@scylladb.com>