sstables: track progress of writing a file
The motivation for that is to track how many partitions were written to a given
file. Tracking this from the caller only without any file_writer help is tricky: calling
write() on the file will keep the writes buffered, and it is not until put() is called
that we have any expectation of that data actually reaching the disk.

However, since the file_writer abstraction is used to wrap the other, non-data
components of the SSTable as well, we will use a general concept of "progress"
instead of partitions. Somebody interested in using this must notify the file_writer that
progress has been made, and when put() is called in the future, that progress is moved,
or commited.

Signed-off-by: Glauber Costa <glauber@scylladb.com>