database: remove unused parameter
We are no longer using the in_flight_seals gate, but forgot to remove it.
To guarantee that all seal operations will have finished when we're done,
we are using the memtable_flush_queue, which also guarantees order. But
that gate was never removed.

The FIXME code should also be removed, since such interface does exist now.

Turns out we will need a gate like that later for the streaming class.
But the name is not ideal anyway, so let's just remove it here, and add
a(nother) gate later when needed.

Signed-off-by: Glauber Costa <glauber@scylladb.com>