repair: rework repair code so we can limit parallelism
The repair code as it is right now is a bit convoluted: it resorts to detached
continuations + do_for_each when calling sync_ranges, and deals with the
problem of excessive parallelism by employing a semaphore inside that range.

Still, even by doing that, we still generate a great number of
checksum requests because the ranges themselves are processed in parallel.

It would be better to just use parallel_for_each for all sub-ranges and let
they all start, and have a per-shard semaphore that limits the overall parallelism
for all requests.

Signed-off-by: Glauber Costa <glauber@scylladb.com>