move throttler inside the memtable list
The main reason we're doing that is that in the presence of many column
families being written to, it can be the case that we'll never issue
CF flushes and deadlock the database.

To fix that, we need to have a way to force the flushing of a column family
into the throttling path. There are two ways of doing that: passing the CF into
the database's throttler, or move the throttler altogether inside a CF member.

This patch takes the second route. The throttler itself still works based on
global parameters (although the waiting list is now local), but will give us
the ability to flush in a natural way within this path.

Signed-off-by: Glauber Costa <glauber@scylladb.com>