memtable list: force flushing before throttling
Throttling happens when the overall shard-wise memory goes over a limit.
However, because the memtables are flushed only when a specific column family's
memtable reach the limit, we can deadlock the database if many CFs are being
actively written to when we reach the global limit.

Imagine the following example: max_memtable_size is 128MB, and
memtable_total_space is 512MB. Now, if only one CF is actively written to this
will never be a problem. But if we have 10 CFs under active writes, we can
reach 512MB by having each of the 10 CFs at 64MB.  In that case, none of them
will flush.

So our criteria here will be to scan all active CFs in the database object,
and flush the largest memory consumer. We'll be careful to keep only one of
those pending_flushes active.

That will be done just for the CF that is using the most amount of memory, and
only once per throttler.

Signed-off-by: Glauber Costa <glauber@scylladb.com>