database: always defer sstable writing.
It is a background task by nature. Basically, if the flush is happening
in the seal_on_overflow() path, we will not be able to return to the client
until we find the first unresolved future. By scheduling a task, we will force
to get the other tasks in the response path back to its running state sooner
rather than later.

Signed-off-by: Glauber Costa <glauber@scylladb.com>