compaction: move everything to a thread
The current code goes through great lengths to mitigate effects of the writer
being run in a seastar thread. I will argue that some of those effects are not
real: since the mutation reader will be effectively called from the writer when
it queries the mutation_reader, that will not, for instance, generate a context
switch.

However, the thing is, we actually want compaction to be run on a thread. All
of it.  The reason is that compaction can get quite hungry, and between two
futures being ready, a lot of processing can ensue. Seastar threads have a
built-in time-based mechanism to yield, and since the writer runs in a thread
anyway, we can just take advantage of that.

Signed-off-by: Glauber Costa <glauber@scylladb.com>