throttler: specify a type for the throttler
We could be throttling writes to either memtables or streaming memtables. We could
detect to which by looking at whether or not we have a parent, but that's way too ad hoc.
Better explicitly pass a type if we need one.

We will need, because in a future patch we will call flush from the throttler.

Signed-off-by: Glauber Costa <glauber@scylladb.com>