sstables: give the code an opportunity to yield
The memtable writer in the SSTable code is used by both the memtable
flush and the compaction processes. In both cases, we need to make sure
that we don't keep running a potentially CPU-intensive loop for a lot of time,
and need to yield instead.

In both cases, we'll yield based on how much time we have been running for.
We'll set to 25 usec. This value is calculated to be the cost of a request
if the shard is doing around 40k requests/sec.

Signed-off-by: Glauber Costa <glauber@scylladb.com>