sstable writer: prevent the creation of a backlog.
If we are CPU-bound, then yielding the memtable/compaction writer will
effectively limit its execution time for only 50usec at every 1msec, or
around 5 % of the CPU time. While this protects the load, it may create
a backlog if the disks are too slow or if the SSTables are too small, or
maybe even both.

We can't allow for that to happen, as a huge number of pending memtables
will throttle the server - which is a less than graceful solution, and a
big compaction backlog may hurt reads.

Because of that, we'll start giving more and more CPU time for the writers
if backlog starts to build.

Signed-off-by: Glauber Costa <glauber@scylladb.com>