force flush of biggest column family
Issue #1195 describes a scenario in which we can freeze the database in the
throttler code.  That would happen when multiple CFs are being written to, and
the sum of their memory usage is larger than the maximum allowed, and yet none
of them is individually over the limit.

The root cause of that is that sealing a memtable is a passive action from the
throttler point of view, only happening when an external factor triggers it,
like an explicit nodetool flush, or a single memtable breaking its maximum.

To fix this problem, we need to also be able to initiate flushes from the
throttling code. With the recent changes that were applied, that becomes
straightforward. The throttler is now visible from the memtable list, and
the memtable list can register itself with it. The throttler will now know
about every memtable list it controls: that is the list of all memtables (or
streaming memtable) for the database.

Once we have that list, it is easy to iterate over it, figure out the biggest
memory user, and trigger a flush. While the reproducer in #1195 would freeze
the database in a matter of seconds, applying this patch allows us to survive
that scenario for 10 minutes without much trouble.

Signed-off-by: Glauber Costa <glauber@scylladb.com>