sstables: enable write behind
This enable write behind for data, index, and the files that go through
write_simple.  The others are less of a concern because they tend to be quite
small, usually smaller than a page. But if this causes no issues or
regressions, we can consider enabling it for them as well.

Signed-off-by: Glauber Costa <glauber@scylladb.com>