serialize memtable flush for a memtable_list
We can only free memory for a region_group when the entire memtable is released.
This means that while the disk can handle requests from multiple memtables just fine,
we won't free any memory until all of them finish. If we are under a pressure situation
we will take a lot more time to leave it.

Since write-behind is now enabled, it's better to serialize the flushes so that only
one memtable is flushed at a time. Having the memtable writer bandwidth all to itself,
the memtable will finish sooner, release memory sooner, and recover the system's health
sooner.

Signed-off-by: Glauber Costa <glauber@scylladb.com>