LSA: enhance eviction function to allow for asynchronous reclaim
This patch augments the eviction_fn object (which used to be just a typedef to a functor), to allow
for both synchronous and asynchronous reclaim.

In doing that, I have explicitly got rid of the call operator overload. While it makes sense to keep
treating this as a function, the fact that behavior now can be synchronous or asynchronous could
confuse users, even if we defaulted to one of them.

Callers will have to explicitly call sync_evict() or async_evict() and should test for the
availability of one or the other.

Signed-off-by: Glauber Costa <glauber@scylladb.com>