LSA: force compaction before commiting to asynchronous reclaim
Specially when we are too close to the throttling point, a single segment that
is freed may be enough to let a request go. We time limited the amount of
compactions we do to make sure we're not dominating the CPU.

This should provide better memory utilization and lower blocking time for
requests.

Signed-off-by: Glauber Costa <glauber@scylladb.com>