LSA: add a virtual destructor to region
Because region is not final, it can be inherited from. As a matter of fact, this
is even likely as the region_group can now tell its users which region is the
largest of them all, and inheritance is a good way to convert the region to the
region memory user. That is, for instance, what we are doing with memtables.

In reality, we could and maybe even should have added this virtual destructor
already since we are doing inheritance in some of our tests. But I have missed
that one. I was reminded about it when g++ complained that region was not a
polymorphic type when trying to cast it back to a memtable.

Signed-off-by: Glauber Costa <glauber@scylladb.com>